fix and clean up
