+++ tests/src/Fixtures/RichEditor/TestRichContentPluginWithoutToolbarButtons.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace Filament\Tests\Fixtures\RichEditor;
+
+use Filament\Actions\Action;
+use Filament\Forms\Components\RichEditor\Plugins\Contracts\RichContentPlugin;
+use Filament\Forms\Components\RichEditor\RichEditorTool;
+use Tiptap\Core\Extension;
+
+class TestRichContentPluginWithoutToolbarButtons implements RichContentPlugin
+{
+    /**
+     * @return array<Extension>
+     */
+    public function getTipTapPhpExtensions(): array
+    {
+        return [];
+    }
+
+    /**
+     * @return array<string>
+     */
+    public function getTipTapJsExtensions(): array
+    {
+        return [];
+    }
+
+    /**
+     * @return array<RichEditorTool>
+     */
+    public function getEditorTools(): array
+    {
+        return [];
+    }
+
+    /**
+     * @return array<Action>
+     */
+    public function getEditorActions(): array
+    {
+        return [];
+    }
+}
