+++ tests/src/Forms/Components/RichEditorTest.php
@@ -1,15 +1,12 @@
 <?php
 
-use Filament\Actions\Action;
 use Filament\Forms\Components\RichEditor;
-use Filament\Forms\Components\RichEditor\Plugins\Contracts\HasToolbarButtons;
-use Filament\Forms\Components\RichEditor\Plugins\Contracts\RichContentPlugin;
-use Filament\Forms\Components\RichEditor\RichEditorTool;
 use Filament\Schemas\Schema;
 use Filament\Tests\Fixtures\Livewire\Livewire;
+use Filament\Tests\Fixtures\RichEditor\TestRichContentPlugin;
+use Filament\Tests\Fixtures\RichEditor\TestRichContentPluginWithoutToolbarButtons;
 use Filament\Tests\TestCase;
 use Illuminate\Validation\ValidationException;
-use Tiptap\Core\Extension;
 
 uses(TestCase::class);
 
@@ -399,50 +396,11 @@ test('`fileAttachments(true)` does not force `attachFiles` button to appear when
 });
 
 test('plugin implementing `HasToolbarButtons` can enable toolbar buttons', function (): void {
-    $plugin = new class implements HasToolbarButtons, RichContentPlugin
-    {
-        /** @return array<Extension> */
-        public function getTipTapPhpExtensions(): array
-        {
-            return [];
-        }
-
-        /** @return array<string> */
-        public function getTipTapJsExtensions(): array
-        {
-            return [];
-        }
-
-        /** @return array<RichEditorTool> */
-        public function getEditorTools(): array
-        {
-            return [];
-        }
-
-        /** @return array<Action> */
-        public function getEditorActions(): array
-        {
-            return [];
-        }
-
-        /** @return array<string | array<string | array<string>>> */
-        public function getEnabledToolbarButtons(): array
-        {
-            return ['highlight'];
-        }
-
-        /** @return array<string> */
-        public function getDisabledToolbarButtons(): array
-        {
-            return [];
-        }
-    };
-
     $richEditor = Schema::make(Livewire::make())
         ->statePath('data')
         ->components([
             RichEditor::make('content')
-                ->plugins([$plugin]),
+                ->plugins([new TestRichContentPlugin(enabledButtons: ['highlight'])]),
         ])
         ->getComponents()[0];
 
@@ -454,50 +412,11 @@ test('plugin implementing `HasToolbarButtons` can enable toolbar buttons', funct
 });
 
 test('plugin implementing `HasToolbarButtons` can disable toolbar buttons', function (): void {
-    $plugin = new class implements HasToolbarButtons, RichContentPlugin
-    {
-        /** @return array<Extension> */
-        public function getTipTapPhpExtensions(): array
-        {
-            return [];
-        }
-
-        /** @return array<string> */
-        public function getTipTapJsExtensions(): array
-        {
-            return [];
-        }
-
-        /** @return array<RichEditorTool> */
-        public function getEditorTools(): array
-        {
-            return [];
-        }
-
-        /** @return array<Action> */
-        public function getEditorActions(): array
-        {
-            return [];
-        }
-
-        /** @return array<string | array<string | array<string>>> */
-        public function getEnabledToolbarButtons(): array
-        {
-            return [];
-        }
-
-        /** @return array<string> */
-        public function getDisabledToolbarButtons(): array
-        {
-            return ['bold', 'italic'];
-        }
-    };
-
     $richEditor = Schema::make(Livewire::make())
         ->statePath('data')
         ->components([
             RichEditor::make('content')
-                ->plugins([$plugin]),
+                ->plugins([new TestRichContentPlugin(disabledButtons: ['bold', 'italic'])]),
         ])
         ->getComponents()[0];
 
@@ -511,50 +430,11 @@ test('plugin implementing `HasToolbarButtons` can disable toolbar buttons', func
 });
 
 test('user `disableToolbarButtons()` overrides plugin-enabled toolbar buttons', function (): void {
-    $plugin = new class implements HasToolbarButtons, RichContentPlugin
-    {
-        /** @return array<Extension> */
-        public function getTipTapPhpExtensions(): array
-        {
-            return [];
-        }
-
-        /** @return array<string> */
-        public function getTipTapJsExtensions(): array
-        {
-            return [];
-        }
-
-        /** @return array<RichEditorTool> */
-        public function getEditorTools(): array
-        {
-            return [];
-        }
-
-        /** @return array<Action> */
-        public function getEditorActions(): array
-        {
-            return [];
-        }
-
-        /** @return array<string | array<string | array<string>>> */
-        public function getEnabledToolbarButtons(): array
-        {
-            return ['highlight'];
-        }
-
-        /** @return array<string> */
-        public function getDisabledToolbarButtons(): array
-        {
-            return [];
-        }
-    };
-
     $schema = Schema::make(Livewire::make())
         ->statePath('data')
         ->components([
             RichEditor::make('content')
-                ->plugins([$plugin]),
+                ->plugins([new TestRichContentPlugin(enabledButtons: ['highlight'])]),
         ]);
 
     $richEditor = $schema->getComponents()[0];
@@ -568,50 +448,11 @@ test('user `disableToolbarButtons()` overrides plugin-enabled toolbar buttons', 
 });
 
 test('user `enableToolbarButtons()` overrides plugin-disabled toolbar buttons', function (): void {
-    $plugin = new class implements HasToolbarButtons, RichContentPlugin
-    {
-        /** @return array<Extension> */
-        public function getTipTapPhpExtensions(): array
-        {
-            return [];
-        }
-
-        /** @return array<string> */
-        public function getTipTapJsExtensions(): array
-        {
-            return [];
-        }
-
-        /** @return array<RichEditorTool> */
-        public function getEditorTools(): array
-        {
-            return [];
-        }
-
-        /** @return array<Action> */
-        public function getEditorActions(): array
-        {
-            return [];
-        }
-
-        /** @return array<string | array<string | array<string>>> */
-        public function getEnabledToolbarButtons(): array
-        {
-            return [];
-        }
-
-        /** @return array<string> */
-        public function getDisabledToolbarButtons(): array
-        {
-            return ['bold'];
-        }
-    };
-
     $schema = Schema::make(Livewire::make())
         ->statePath('data')
         ->components([
             RichEditor::make('content')
-                ->plugins([$plugin]),
+                ->plugins([new TestRichContentPlugin(disabledButtons: ['bold'])]),
         ]);
 
     $richEditor = $schema->getComponents()[0];
@@ -624,38 +465,11 @@ test('user `enableToolbarButtons()` overrides plugin-disabled toolbar buttons', 
 });
 
 test('plugin without `HasToolbarButtons` does not affect toolbar buttons', function (): void {
-    $plugin = new class implements RichContentPlugin
-    {
-        /** @return array<Extension> */
-        public function getTipTapPhpExtensions(): array
-        {
-            return [];
-        }
-
-        /** @return array<string> */
-        public function getTipTapJsExtensions(): array
-        {
-            return [];
-        }
-
-        /** @return array<RichEditorTool> */
-        public function getEditorTools(): array
-        {
-            return [];
-        }
-
-        /** @return array<Action> */
-        public function getEditorActions(): array
-        {
-            return [];
-        }
-    };
-
     $richEditor = Schema::make(Livewire::make())
         ->statePath('data')
         ->components([
             RichEditor::make('content')
-                ->plugins([$plugin]),
+                ->plugins([new TestRichContentPluginWithoutToolbarButtons]),
         ])
         ->getComponents()[0];
